+++ app/Tests/RequirementTests/ModifierTests.hs
@@ -5,7 +5,7 @@ Module containing test cases for Modifier string formatters.
 
 -}
 
-module RequirementTests.ModifierTests
+module Tests.RequirementTests.ModifierTests
 ( modifierTestSuite ) where
 
 import Database.Requirement